move chartData into server/component folder
